Separate Java 9 test
