+++ core/build.gradle
@@ -13,12 +13,29 @@ sourceSets {
             srcDirs = ['src/main/java9']
         }
     }
+    java9Test {
+        java {
+            srcDirs = ['src/test/java9']
+        }
+    }
 }
 
 compileJava9Java {
     sourceCompatibility = 9
     targetCompatibility = 9
 }
+
+compileJava9TestJava {
+    sourceCompatibility = 9
+    targetCompatibility = 9
+}
+
+task testJava9(type: Test, group: 'Verification', description: 'Runs unit tests for Java 9 source set') {
+    testClassesDirs = sourceSets.java9Test.output.classesDirs
+    classpath = sourceSets.java9Test.runtimeClasspath
+}
+
+check.dependsOn testJava9
 
 tasks.withType(JavaCompile) {
     def version = releaseFlagFromVersion(sourceCompatibility)
@@ -53,6 +70,8 @@ dependencies {
     // Provide the dependencies of the main sourceset to others. Do not ever add version-specific dependencies
     // to a sourceset, it won't work.
     java9Implementation sourceSets.main.compileClasspath
+    java9TestImplementation sourceSets.java9.compileClasspath
+    java9TestImplementation sourceSets.test.compileClasspath
 
     // cglib
     testCompile 'cglib:cglib'
